# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API",
        "description": "API Documentation",
        "version": "1.0.0"
    },
    "paths": {
        "/api/genres": {
            "get": {
                "tags": [
                    "Local Api"
                ],
                "summary": "Get All Genres",
                "operationId": "dc264da58b11271402354527a5f68d3f",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/artist": {
            "get": {
                "tags": [
                    "Local Api"
                ],
                "summary": "Get Artist",
                "description": "Required => token, artistId",
                "operationId": "e088c6d3ded7fa0b1470dc9b5ccd71ad",
                "parameters": [
                    {
                        "name": "artistId",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/artist/album": {
            "get": {
                "tags": [
                    "Local Api"
                ],
                "summary": "Get Artist's Album",
                "description": "Required => token, artistId",
                "operationId": "6c76143ed3cd665757b054c8e50e55fb",
                "parameters": [
                    {
                        "name": "artistId",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/artist/tracks": {
            "get": {
                "tags": [
                    "Local Api"
                ],
                "summary": "Get Artist's Tracks",
                "description": "Required => token, artistId",
                "operationId": "95d2faf800ade4a56e060ded4d77d629",
                "parameters": [
                    {
                        "name": "artistId",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/album": {
            "get": {
                "tags": [
                    "Local Api"
                ],
                "summary": "Get Album",
                "description": "Required => token, albumId",
                "operationId": "1f38ce1c1b54f3ae6421bacca3a4405f",
                "parameters": [
                    {
                        "name": "albumId",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "Bearer": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}